suncc fixes

+++ src/engine/kernel/api/kernel/suncc/interlocked.inl
@@ -188,8 +188,8 @@ struct InterlockedType<8>
     }
     static inline bool set_conditional(tagged_t *p, tagged_t::value_t v, tagged_t::tag_t& condition)
     {
-        return set_conditional_128((i64*)p, v, condition.taggedvalue.tag,
-                                   condition.taggedvalue.value);
+        return set_conditional_128((i64*)p, (i64)v, condition.taggedvalue.tag,
+                                   (i64)condition.taggedvalue.value);
     }
 };
 #endif
